+++ src/main/java/com/jiro4989/mvte/table/TextTable.java
@@ -5,6 +5,7 @@ import static com.jiro4989.mvte.util.Utils.r;
 
 import com.jiro4989.mvte.MainController;
 import com.jiro4989.mvte.selector.ImageSelector;
+import com.jiro4989.mvte.util.Brackets;
 import com.jiro4989.mvte.util.JsonUtils;
 import java.io.*;
 import java.util.*;
@@ -16,7 +17,6 @@ import javafx.scene.control.*;
 import javafx.scene.control.cell.*;
 import javafx.scene.input.*;
 import javax.xml.parsers.ParserConfigurationException;
-import jiro.java.lang.Brackets;
 import jiro.java.util.MyProperties;
 
 public class TextTable {
