+++ src/main/java/com/jiro4989/mvte/util/ActorNameType.java
@@ -1,4 +1,4 @@
-package jiro.java.lang;
+package com.jiro4989.mvte.util;
 
 /** アクター名の表示スタイル列挙型 */
 public enum ActorNameType {
